Changed chunk saving to unloading events
To avoid the 'empty chunks on load' issue, set the save operation to
only work when chunks are unloaded rather than loaded/unloaded. This
also means the mod does half as much work, since it doesn't store the
chunk on both load and unload.

This meant that chunks that had not been unloaded yet would not be
saved. Added a step to the shutdown method so that currently loaded
chunks are saved first. This also means chunks that have loaded but not
yet unloaded will also be stored.

Changed the save file name to incorporate the world name. While the
timestamp in the file name makes it unlikely you will get a collision,
it would have been a pain to sort out which files apply to which server
if you played on multiple servers.