Improvements and Bug Fixes
The mod now trims chunks down to the nearest 500 after you reach 10,000.
Tested on my Java - the file writer is asynchronous so this shouldn't
cause too much lag when it fires.

Fixed a bug where disconnecting from a server would write the files to
disk and close the writer, but wouldn't clear the stored chunks from
memory. This caused the chunks to be written back to disk again even if
they weren't updated, making for redundant data.

Cleaned up the mod so that file writing actions are part of a writer
object and not the core of the mod. Also took the Forge hooks out of the
core and put them in a generic mod_GeoSharer file, which should make it
easier to implement in other mod environments by changing those hooks.
This also makes MagicLauncher recognise the .zip as an external mod.

The merge utility now has a very simple save process incorporated into
it. In the command line enter "output [directory]" and "input
[path\filesearch]" and that will choose the respective files. The input
can use wildcards for the filename, which will select all the files
matching the search and do them in sequence. The utility won't start
working until you issue a "run" command, giving the user the chance to
change the settings.

The .geosharer file format is now changed. Block data is now squashed
into 4-bits, with each byte holding the block data for 2 blocks. The
"air only" layers at the top of a chunk are not saved at all, which
saves storage space and marginally increases merge speed. The final
output is saved with an additional layer of GZip compression, which
saves more space while still allowing for the "ReadLine" function in the
stream reader. This brings the average size of a chunk down to ~1KB.

Note that this change in format has broken the ability of the merge
utility to process older files, but since nobody was using the mod
anyway that shouldn't be a big problem.

There were also some minor tweaks to the merge process to try and speed
it up a little. Unfortunately there is a limit to what can be done since
block relighting and fluid dynamics are intensive processes.