Multi file reader for merge utility
Changed the file reader class to have its enumerator as a nested class.
Also changed the reader to accept multiple input files, and the
enumerator to be able to enumerate through all files from start to
finish. Fixed a few break conditions, such as on null chunks, so that
they wouldn't prematurely abort the entire process.

Also amended the feedback text in console so that expected errors are
not reported, and progress updates are given on every chunk. The process
updates are passed off to another thread so that it doesn't interfere
with the merge process.