Snapshotting and DragDrop
Added overloads to GeoMultifile's GetLatest methods to take a DateTime
value representing the latest chunk timestamps to accept. This allows
for creating a snapshot of the world up to a given datetime.

Also added drag+drop functionality to the winform. Drop a single
directory to set the output directory, or drop any number of *.geosharer
files to add them to the list of input files.